+++ app/graphql/types/project_type.rb
@@ -105,5 +105,6 @@ module Types
     field :gis_url, String, null: true
     field :info_manager_url, String, null: true
     field :confirmation_status, String, null: true
+    field :strategic_partner, String, null: true
   end
 end