+++ db/migrate/20220503111729_add_strategic_partner_to_projects.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddStrategicPartnerToProjects < ActiveRecord::Migration[6.1]
+  def change
+    safety_assured { add_column :projects, :strategic_partner, :string }
+  end
+end